chore(#12): dead_code cleanup + log redaction + cargo audit CI (#13)
* chore(quality): address #12 low-priority items — dead_code cleanup, log redaction, cargo audit

PR #5 引き継ぎ項目:
- 各 crate の crate-wide `#![allow(dead_code, unused_imports, unused_variables)]`
  を除去し、必要な箇所は struct / enum / impl レベルの個別 `#[allow(dead_code)]`
  (将来用途のため残す API に限る) に置換。
- `ProjectManager::open/close/list` の後方互換 thin wrapper を削除
  (trait `ProjectConfiguration` の `open_project/close_project/list_projects`
  と完全重複していた)。`close_all` は Daemon shutdown 経路で使用中なので残置。
- `cargo fix` で unused imports を自動除去。

ログ露出削減 (S24 一部):
- `PeerId::short()` を追加: 先頭 12 文字 + `…` の短縮表示。
- conduit/quic の `tracing::info!("... peer {} ...")` を `peer_id.short()` に
  差し替え (4 箇所)。全箇所への波及は次フェーズで。

CI 強化:
- `.github/workflows/ci.yml` に `cargo audit` job を追加
  (`--deny unsound --deny yanked` で unsound + yanked クレートをブロック。
  unmaintained 警告は当面許容)。

検証: cargo fmt --all / clippy -D warnings / test --all (34/34 pass)。

* fix(ci,ipc): restore Duration import on unix + non-blocking cargo audit

- synergos-core/ipc_server.rs: `cargo fix` が Windows ビルドしか見ずに
  `std::time::Duration` の use を削除してしまっていた。`#[cfg(unix)]` で
  再度 import する (Unix 側の tokio::time::sleep(Duration::...) が使う)。
- .github/workflows/ci.yml: cargo audit を `continue-on-error: true` に。
  rustls / quinn / hickory-resolver などの transitive dep に RUSTSEC-2026
  系の脆弱性が複数あり、これらは upstream 追従でしか直せないので CI
  ブロックしない運用にする。情報提供として毎回実行する。

+++ synergos-net/src/conduit/mod.rs
@@ -8,11 +8,10 @@ use std::time::{Duration, Instant};
 
 use dashmap::DashMap;
 
-use crate::config::NetConfig;
 use crate::error::{Result, SynergosNetError};
 use crate::mesh::{Mesh, ProbeResult};
-use crate::quic::{QuicConnectionInfo, QuicManager};
-use crate::tunnel::{TunnelManager, TunnelState};
+use crate::quic::QuicManager;
+use crate::tunnel::TunnelManager;
 use crate::types::{ConnectionState, PeerEndpoint, PeerId, Route, RouteKind};
 
 /// 経路検出結果
@@ -50,6 +49,7 @@ struct ManagedPeer {
 ///
 /// 各ピアへの最適な接続経路を選択し、接続の確立・維持・切り替えを管理する。
 /// IPv6 Direct → Tunnel → Relay の優先度で接続を試行する。
+#[allow(dead_code)]
 pub struct Conduit {
     /// 管理中のピア（PeerId → ManagedPeer）
     peers: DashMap<PeerId, ManagedPeer>,
@@ -144,7 +144,7 @@ impl Conduit {
                     };
                     entry.last_seen = Instant::now();
                 }
-                tracing::info!("Connected to peer {} via {:?}", peer_id, route_kind);
+                tracing::info!("Connected to peer {} via {:?}", peer_id.short(), route_kind);
                 Ok(route_kind)
             }
             Err(e) => {
@@ -170,7 +170,7 @@ impl Conduit {
             entry.active_route = None;
         }
 
-        tracing::info!("Disconnected peer {}: {}", peer_id, reason);
+        tracing::info!("Disconnected peer {}: {}", peer_id.short(), reason);
         Ok(())
     }
